+++ src/Chunkyard/Cli/Commands.cs
@@ -107,18 +107,10 @@ namespace Chunkyard.Cli
                 return new FileStream(file, mode, FileAccess.Write);
             }
 
-            var restoredBlobs = snapshotStore.RestoreSnapshot(
+            snapshotStore.RestoreSnapshot(
                 o.SnapshotId,
                 new Fuzzy(o.IncludePatterns, emptyMatches: true),
                 OpenWrite);
-
-            foreach (var blob in restoredBlobs)
-            {
-                var path = Path.Combine(o.Directory, blob.Name);
-
-                File.SetCreationTimeUtc(path, blob.CreationTimeUtc);
-                File.SetLastWriteTimeUtc(path, blob.LastWriteTimeUtc);
-            }
         }
 
         public static void ListSnapshots(ListOptions o)